requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-environ==0.11.2
django-filter==24.2
celery==5.3.6
redis==5.0.4
django-redis==5.4.0
Pillow==10.3.0